+++ src/misc/coders/Encoders/video/h264_libx264.js
@@ -2,7 +2,6 @@ import React from 'react';
 
 import Grid from '@mui/material/Grid';
 import MenuItem from '@mui/material/MenuItem';
-import TextField from '@mui/material/TextField';
 
 import { Trans } from '@lingui/macro';
 
@@ -12,12 +11,13 @@ import Helper from '../../helper';
 
 function init(initialState) {
 	const state = {
+		preset: 'ultrafast',
 		bitrate: '4096',
 		fps: '25',
 		gop: '2',
-		profile: '77',
-		rc_mode: '1',
-		quality: '-1',
+		profile: 'auto',
+		tune: 'zerolatency',
+		fps_mode: 'auto',
 		...initialState,
 	};
 
@@ -28,73 +28,90 @@ function createMapping(settings, stream, skills) {
 	stream = Helper.InitStream(stream);
 	skills = Helper.InitSkills(skills);
 
-	const global = [];
-	const local = [];
-
-	// https://trac.ffmpeg.org/wiki/Hardware/VAAPI
-	global.push(['-vaapi_device', '/dev/dri/renderD128']);
-
-	local.push(
-		'-vf',
-		'format=nv12,hwupload',
+	const local = [
 		'-codec:v',
-		'hevc_vaapi',
-		'-profile:v',
-		`${settings.profile}`,
-		'-quality',
-		`${settings.quality}`,
+		'libx264',
+		'-preset:v',
+		`${settings.preset}`,
 		'-b:v',
 		`${settings.bitrate}k`,
-		'-maxrate',
+		'-maxrate:v',
 		`${settings.bitrate}k`,
-		'-bufsize',
+		'-bufsize:v',
 		`${settings.bitrate}k`,
 		'-r',
 		`${settings.fps}`,
-		'-g',
-		`${settings.gop}`
-	);
+		'-sc_threshold',
+		'0',
+		'-pix_fmt',
+		'yuv420p',
+	];
 
 	if (settings.gop !== 'auto') {
-		local.push('-g', `${Math.round(parseInt(settings.fps) * parseInt(settings.gop)).toFixed(0)}`);
+		local.push(
+			'-g',
+			`${Math.round(parseInt(settings.fps) * parseInt(settings.gop)).toFixed(0)}`,
+			'-keyint_min',
+			`${Math.round(parseInt(settings.fps) * parseInt(settings.gop)).toFixed(0)}`,
+		);
 	}
 
-	return {
-		global: global,
+	if (skills.ffmpeg.version_major >= 5) {
+		local.push('-fps_mode', `${settings.fps_mode}`);
+	}
+
+	if (settings.profile !== 'auto') {
+		local.push('-profile:v', `${settings.profile}`);
+	}
+
+	if (settings.tune !== 'none') {
+		local.push('-tune:v', `${settings.tune}`);
+	}
+
+	const mapping = {
+		global: [],
 		local: local,
 	};
+
+	return mapping;
 }
 
-function RateControl(props) {
+function Preset(props) {
 	return (
-		<Select label={<Trans>Rate control</Trans>} value={props.value} onChange={props.onChange}>
-			<MenuItem value="0">auto</MenuItem>
-			<MenuItem value="1">Constant-quality</MenuItem>
-			<MenuItem value="2">Constant-bitrate</MenuItem>
-			<MenuItem value="3">Variable-bitrate</MenuItem>
-			<MenuItem value="4">Intelligent constant-quality</MenuItem>
-			<MenuItem value="5">Quality-defined variable-bitrate</MenuItem>
-			<MenuItem value="6">Average variable-bitrate</MenuItem>
+		<Select label={<Trans>Preset</Trans>} value={props.value} onChange={props.onChange}>
+			<MenuItem value="ultrafast">ultrafast</MenuItem>
+			<MenuItem value="superfast">superfast</MenuItem>
+			<MenuItem value="veryfast">veryfast</MenuItem>
+			<MenuItem value="faster">faster</MenuItem>
+			<MenuItem value="fast">fast</MenuItem>
+			<MenuItem value="medium">medium</MenuItem>
+			<MenuItem value="slow">slow</MenuItem>
+			<MenuItem value="slower">slower</MenuItem>
+			<MenuItem value="veryslow">veryslow</MenuItem>
 		</Select>
 	);
 }
 
-RateControl.defaultProps = {
+Preset.defaultProps = {
 	value: '',
 	onChange: function (event) {},
 };
 
-function Profile(props) {
+function Tune(props) {
 	return (
-		<Select label={<Trans>Profile</Trans>} value={props.value} onChange={props.onChange}>
-			<MenuItem value="578">baseline (constrained)</MenuItem>
-			<MenuItem value="77">main</MenuItem>
-			<MenuItem value="100">high</MenuItem>
+		<Select label={<Trans>Tune</Trans>} value={props.value} onChange={props.onChange}>
+			<MenuItem value="none">none</MenuItem>
+			<MenuItem value="animation">animation</MenuItem>
+			<MenuItem value="fastdecode">fastdecode</MenuItem>
+			<MenuItem value="film">film</MenuItem>
+			<MenuItem value="grain">grain</MenuItem>
+			<MenuItem value="stillimage">stillimage</MenuItem>
+			<MenuItem value="zerolatency">zerolatency</MenuItem>
 		</Select>
 	);
 }
 
-Profile.defaultProps = {
+Tune.defaultProps = {
 	value: '',
 	onChange: function (event) {},
 };
@@ -131,9 +148,6 @@ function Coder(props) {
 	return (
 		<Grid container spacing={2}>
 			<Grid item xs={12}>
-				<RateControl value={settings.rc_mode} onChange={update('rc_mode')} />
-			</Grid>
-			<Grid item xs={12} md={6}>
 				<Video.Bitrate value={settings.bitrate} onChange={update('bitrate')} allowCustom />
 			</Grid>
 			<Grid item xs={12} md={6}>
@@ -142,11 +156,19 @@ function Coder(props) {
 			<Grid item xs={12} md={6}>
 				<Video.GOP value={settings.gop} onChange={update('gop')} allowAuto allowCustom />
 			</Grid>
+			{skills.ffmpeg.version_major >= 5 && (
+				<Grid item xs={12}>
+					<Video.FpsMode value={settings.fps_mode} onChange={update('fps_mode')} />
+				</Grid>
+			)}
 			<Grid item xs={6}>
-				<TextField variant="outlined" fullWidth label={<Trans>Quality</Trans>} type="number" value={settings.quality} onChange={update('quality')} />
+				<Preset value={settings.preset} onChange={update('preset')} />
 			</Grid>
 			<Grid item xs={6}>
-				<Profile value={settings.profile} onChange={update('profile')} />
+				<Video.Profile value={settings.profile} onChange={update('profile')} />
+			</Grid>
+			<Grid item xs={6}>
+				<Tune value={settings.tune} onChange={update('tune')} />
 			</Grid>
 		</Grid>
 	);
@@ -159,14 +181,14 @@ Coder.defaultProps = {
 	onChange: function (settings, mapping) {},
 };
 
-const coder = 'hevc_vaapi';
-const name = 'HEVC (Intel VAAPI)';
-const codec = 'hevc';
+const coder = 'libx264';
+const name = 'H.264 (libx264)';
+const codec = 'h264';
 const type = 'video';
-const hwaccel = true;
+const hwaccel = false;
 
 function summarize(settings) {
-	return `${name}, ${settings.bitrate} kbit/s, ${settings.fps} FPS, Profile: ${settings.profile}`;
+	return `${name}, ${settings.bitrate} kbit/s, ${settings.fps} FPS, Preset: ${settings.preset}, Profile: ${settings.profile}`;
 }
 
 function defaults(stream, skills) {